update wiremock stubs and split tests into api and application tests
