+++ src/carbonscore/settings.py
@@ -28,7 +28,7 @@ SECRET_KEY = os.environ.get('SECRET_KEY')
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = os.environ.get('DEBUG') == 'True'
 
-ALLOWED_HOSTS = ['127.0.0.1', 'localhost', 'sqs-application-web-1']
+ALLOWED_HOSTS = ['127.0.0.1', 'localhost', 'sqs-application-web-1', 'testserver']
 
 
 # Application definition
