+++ wiremock/stubs/__files/auth.json
@@ -1,0 +1,3 @@
+{
+    "message":"Authentication successful"
+}
