+++ wiremock/stubs/__files/electricity.json
@@ -1,0 +1,1 @@
+{"data": {"id": "ef065a14-651f-4323-a64a-83890348af22","type": "estimate","attributes": {"country": "us","state": "fl","electricity_unit": "kwh","electricity_value": 123.45,"estimated_at": "2024-05-09T17:25:52.041Z","carbon_g": 48937,"carbon_lb": 107.89,"carbon_kg": 48.94,"carbon_mt": 0.05}}}
