+++ wiremock/stubs/__files/fuel.json
@@ -1,0 +1,1 @@
+{"data": {"id": "97e1b97b-9c3b-4e87-b2f4-4c20131d162d","type": "estimate","attributes": {"fuel_source_type": "ng","fuel_source_unit": "thousand_cubic_feet","fuel_source_value": 500.0,"estimated_at": "2024-05-09T17:34:39.806Z","carbon_g": 26560000,"carbon_lb": 58554.78,"carbon_kg": 26560.0,"carbon_mt": 26.56}}}
