+++ wiremock/stubs/__files/shipping.json
@@ -1,0 +1,1 @@
+{"data": {"id": "b4ac6ce9-ba1b-4ee4-b30e-b73e16465bfd","type": "estimate","attributes": {"distance_value": 500.01,"weight_unit": "g","transport_method": "plane","weight_value": 123.45,"distance_unit": "km","estimated_at": "2024-05-09T17:18:01.418Z","carbon_g": 37,"carbon_lb": 0.08,"carbon_kg": 0.04,"carbon_mt": 0.0}}}
